% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{getClimateData}
\alias{getClimateData}
\title{Access the climate change data.}
\usage{
getClimateData(modelID, modelFrequency, modelVar, lon, lat, iMonth = 1, iYear,
  fMonth = 1, fYear)
}
\value{
TESTE.
}
\description{
\code{getClimateData} access the climate change data from CPTEC/INPE.
}
\examples{
getClimateData('1', 'YEARLY','TP2M', '-12', '-49', iYear = 2006, fYear = 2010)
}
