% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{getClimateDataBR}
\alias{getClimateDataBR}
\title{Access the Brazil climate change data.}
\usage{
getClimateDataBR(modelID, modelFrequency, modelVar, iYear, fYear)
}
\arguments{
\item{modelID}{numeric (model ID).}

\item{modelFrequency}{string (data frequency).}

\item{modelVar}{string (model variable short name).}

\item{iYear}{numeric (initial year).}

\item{fYear}{numeric (final year).}
}
\value{
list (list with climate change data)
}
\description{
\code{getClimateDataBR} access the Brazil climate change data from CPTEC/INPE.
}
\examples{
\dontrun{
getClimateDataBR('1', 'YEARLY', 'TP2M', iYear = 2006, fYear = 2006)
}
}
