% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotGrafData}
\alias{plotGrafData}
\title{Create a graph with climate change data.}
\usage{
plotGrafData(modelID, modelFrequency, modelVar, lat, lon, iYear, fYear)
}
\arguments{
\item{modelID}{numeric (model ID).}

\item{modelFrequency}{string (data frequency).}

\item{modelVar}{string (model variable short name).}

\item{lat}{numeric (latitude coordenate).}

\item{lon}{numeric (longitude coordenate).}

\item{iYear}{numeric (initial year).}

\item{fYear}{numeric (final year).}
}
\value{
graph (graph with climate change data)
}
\description{
\code{plotGrafData} create a graph with climate change data from CPTEC/INPE.
}
\examples{
\dontrun{
 plotGrafData('1', 'MONTHLY', 'TP2M', '-28.35', '-52.34', iYear = 2006, fYear = 2010)
}
}
