% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotMapBR}
\alias{plotMapBR}
\title{Create a map with Brazil climate change data.}
\usage{
plotMapBR(modelID, modelVar, year)
}
\arguments{
\item{modelID}{numeric (model ID).}

\item{modelVar}{string (model variable short name).}

\item{year}{numeric (year).}
}
\value{
map (map with climate change data)
}
\description{
\code{plotMapBR} create a map with Brazil climate change data from CPTEC/INPE.
}
\examples{
\dontrun{
 plotMapBR('1', 'PREC', 2006)
 plotMapBR('1', 'TP2M', 2006)
}
}
