% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotMapBRgg}
\alias{plotMapBRgg}
\title{Create a map with Brazil climate change data.}
\usage{
plotMapBRgg(modelID, modelVar, year)
}
\arguments{
\item{modelID}{numeric (model ID).}

\item{modelVar}{string (model variable short name).}

\item{year}{numeric (year).}
}
\value{
map (map with climate change data)
}
\description{
\code{plotMapBRgg} create a map with Brazil climate change data from CPTEC/INPE using ggplot2.
}
\examples{
\dontrun{
 plotMapBRgg('1', 'PREC', 2006)
 plotMapBRgg('1', 'TP2M', 2006)
}
}
