% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotMapPontos}
\alias{plotMapPontos}
\title{Create a map with the climate change data from the rectangular area between 2 points.}
\usage{
plotMapPontos(modelID, modelVar, lat1, lon1, lat2, lon2, year)
}
\arguments{
\item{modelID}{numeric (model ID).}

\item{modelVar}{string (model variable short name).}

\item{lat1}{numeric (latitude coordenate).}

\item{lon1}{numeric (longitude coordenate).}

\item{lat2}{numeric (latitude coordenate).}

\item{lon2}{numeric (longitude coordenate).}

\item{year}{numeric (year).}
}
\value{
map (map with climate change data from the rectangular area between 2 points)
}
\description{
\code{plotMapPontos} create a map with the climate change data from the rectangular area between 2 points.
}
\examples{
\dontrun{
 plotMapPontos('2', 'PREC','-27.26', '-57.10', '-33.67', '-48.85', year = 2006)
 plotMapPontos('2', 'PREC', '-35.05','-23.95','5.9','-75.05', year = 2006)
}
}
