% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotMapShape}
\alias{plotMapShape}
\title{Create a map with the climate data from the shapefile area.}
\usage{
plotMapShape(modelID, modelVar, year, folderPath, fileName,
  subName = NULL, subNameValue = NULL)
}
\arguments{
\item{modelID}{numeric (model ID).}

\item{modelVar}{string (model variable short name).}

\item{year}{numeric (year).}

\item{folderPath}{xxx}

\item{fileName}{xxx}

\item{subName}{xx}

\item{subNameValue}{xxx}
}
\value{
map (map with climate change data from the rectangular area between 2 points)
}
\description{
\code{plotMapShape} create a map with the climate data from the shapefile area.
}
\examples{
\dontrun{
  plotMapShape('2', 'PREC', 2006, "C:/path/to/your/folder", "FileNameWithoutExtension")
  plotMapShape('1', 'PREC', 2006, "C:/Users/Lemon/Desktop/Shapefile", "teste", "sigla", "RS")
  plotMapShape('1', 'PREC', 2006, "C:/Users/Lemon/Desktop/Shapefile", "teste", "regiao_id", "3")
}
}
