% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{ProjetaVisual}
\alias{ProjetaVisual}
\title{Create a app shiny to access data from Projeta database.}
\usage{
ProjetaVisual()
}
\description{
\code{ProjetaVisual} app shiny to access data from Projeta database.
}
\examples{
\dontrun{
  ProjetaVisual()
}
}
