% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{getClimateData}
\alias{getClimateData}
\title{Access the climate change data.}
\usage{
getClimateData(modelID, modelFrequency, modelVar, lat, lon, iYear, fYear)
}
\arguments{
\item{modelID}{numeric (model ID).}

\item{modelFrequency}{string (data frequency).}

\item{modelVar}{string (model variable short name).}

\item{lat}{numeric (latitude coordenate).}

\item{lon}{numeric (longitude coordenate).}

\item{iYear}{numeric (initial year).}

\item{fYear}{numeric (final year).}
}
\value{
list (list with climate change data)
}
\description{
\code{getClimateData} access the climate change data from CPTEC/INPE.
}
\examples{
getClimateData('1', 'YEARLY', 'TP2M', '-28.35', '-52.34', iYear = 2006, fYear = 2010)
}
