% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R, R/visualization.R
\name{getInfoClimate}
\alias{getInfoClimate}
\title{Information about the climate change data.}
\usage{
getInfoClimate()

getInfoClimate()
}
\value{
Model driven, frequencies and variables.

Model driven, frequencies and variables.
}
\description{
\code{getInfoClimate} returns information about the climate change data accessed from CPTEC/INPE.

\code{getInfoClimate} returns information about the climate change data accessed from CPTEC/INPE.
}
\examples{
getInfoClimate()
getInfoClimate()
}
